Add add-delete test
