+++ tests/blog-posts/add-update.test.js
@@ -7,24 +7,6 @@ describe('Add update blog posts', function () {
 
   let blogPost;
 
-  const newPosts = [
-    {
-      title: 'New Title 2',
-      body: 'New blog details 2',
-      userId: 2,
-    },
-    {
-      title: 'New Title 3',
-      body: 'New blog details 3',
-      userId: 3,
-    },
-    {
-      title: 'New Title 4',
-      body: 'New blog details 4',
-      userId: 4,
-    },
-  ];
-
   before('Get a blog post to use for searching and updates', async function () {
     await blogPosts.getAllPosts().then((response) => {
       expect(response.status).to.equal(StatusCode.Ok);
@@ -68,23 +50,6 @@ describe('Add update blog posts', function () {
 
       await blogPosts.updatePost(postUpdate).then((response) => {
         expect(response.status).to.equal(StatusCode.InternalServerError);
-      });
-    });
-  });
-
-  describe.only('Create and remove multiple blog posts one after another', function () {
-    newPosts.forEach((newPost) => {
-      it(`should add a new blog post then delete it: ${newPost.title}`, async function () {
-        await blogPosts.addPost(newPost).then((response) => {
-          expect(response.status).to.equal(StatusCode.Created);
-          expect(response.data).to.have.property('id');
-        });
-        await blogPosts.getPost(newPost.userId).then((response) => {
-          expect(response.status).to.equal(StatusCode.Ok);
-        });
-        await blogPosts.deletePost(blogPost.id).then((response) => {
-          expect(response.status).to.equal(StatusCode.Ok);
-        });
       });
     });
   });
